pyproject.toml
[project]
name = "ecommerce-analitic-service"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = []

[dependency-groups]
dev = [
    "asgiref==3.10.0",
    "django==5.2.7",
    "djangorestframework==3.16.1",
    "gunicorn==23.0.0",
    "packaging==25.0",
    "psycopg2-binary==2.9.11",
    "python-dotenv==1.1.1",
    "sqlparse==0.5.3",
    "tzdata==2025.2",
]
